Restructure module for adding new structures.py
- Add akd/_base/struvtures.py
